Move existing helper to util file
