+++ packages/back-end/src/api/features/postFeature.ts
@@ -2,9 +2,7 @@ import { z } from "zod";
 import { validateFeatureValue, validateScheduleRules } from "shared/util";
 import { PostFeatureResponse } from "shared/types/openapi";
 import { postFeatureValidator } from "shared/validators";
-import { FeatureInterface, JSONSchemaDef } from "shared/types/feature";
-import { OrganizationInterface } from "shared/types/organization";
-import { orgHasPremiumFeature } from "back-end/src/enterprise";
+import { FeatureInterface } from "shared/types/feature";
 import { createApiRequestHandler } from "back-end/src/util/handler";
 import { createFeature, getFeature } from "back-end/src/models/FeatureModel";
 import { getExperimentMapForFeature } from "back-end/src/models/ExperimentModel";
@@ -19,7 +17,7 @@ import { auditDetailsCreate } from "back-end/src/services/audit";
 import { getEnvironments } from "back-end/src/services/organizations";
 import { getRevision } from "back-end/src/models/FeatureRevisionModel";
 import { addTags } from "back-end/src/models/TagModel";
-import { logger } from "back-end/src/util/logger";
+import { parseApiJsonSchema } from "back-end/src/util/feature-json-schema";
 import { validateCustomFields } from "./validations";
 
 export type ApiFeatureEnvSettings = NonNullable<
@@ -41,30 +39,6 @@ export const validateEnvKeys = (
         "', '",
       )}' not recognized. Please create the environment or remove it from your environment settings and try again.`,
     );
-  }
-};
-
-export const parseJsonSchemaForEnterprise = (
-  org: OrganizationInterface,
-  jsonSchema: string | undefined,
-) => {
-  const jsonSchemaWrapper: JSONSchemaDef = {
-    schemaType: "schema",
-    schema: "",
-    simple: { type: "object", fields: [] },
-    date: new Date(),
-    enabled: false,
-  };
-  if (!jsonSchema) return jsonSchemaWrapper;
-  if (!orgHasPremiumFeature(org, "json-validation")) return jsonSchemaWrapper;
-  try {
-    // ensure the schema is valid JSON
-    jsonSchemaWrapper.schema = JSON.stringify(JSON.parse(jsonSchema));
-    jsonSchemaWrapper.enabled = true;
-    return jsonSchemaWrapper;
-  } catch (e) {
-    logger.error(e, "Failed to parse feature json schema");
-    return jsonSchemaWrapper;
   }
 };
 
@@ -177,10 +151,7 @@ export const postFeature = createApiRequestHandler(postFeatureValidator)(async (
 
   feature.environmentSettings = environmentSettings;
 
-  const jsonSchema = parseJsonSchemaForEnterprise(
-    req.context.org,
-    req.body.jsonSchema,
-  );
+  const jsonSchema = parseApiJsonSchema(req.context.org, req.body.jsonSchema);
 
   feature.jsonSchema = jsonSchema;
